(tests): Minor refactor to prevent duplication
